backend/huggingface-deploy/requirements.txt
fastapi
uvicorn[standard]
pillow
python-multipart
ultralytics
torch
torchvision
numpy
opencv-python-headless
